// web/generated/schema.ts
export interface Menu extends SanityDocument {
  _type: "menu"

  /**
   * Title — `string`
   *
   *
   */
  title?: string

  /**
   * Manu Items — `array`
   *
   *
   */
  items?: Array<SanityKeyedReference<Page>>
}

/**
 * Page
 *
 *
 */
export interface Page extends SanityDocument {
  _type: "page"

  /**
   * Title — `localeString`
   *
   *
   */
  title?: LocaleString

  /**
   * Menu Title — `localeString`
   *
   * The title shown in the main navigation
   */
  menuTitle?: LocaleString

  /**
   * Subtitle — `localeString`
   *
   * Not available for 'Page' type
   */
  subtitle?: LocaleString

  /**
   * Template — `array`
   *
   *
   */
  template?: Array<SanityKeyed<string>>

  /**
   * Slug — `localeSlug`
   *
   *
   */
  slug?: LocaleSlug

  /**
   * Body — `localeRichText`
   *
   * Only available for 'Page' type
   */
  body?: LocaleRichText

  /**
   * Main image — `captionImage`
   *
   * Only available for 'Page' type
   */
  mainImage?: CaptionImage

  /**
   * SEO title — `localeString`
   *
   * Displayed on Facebook and Twitter shares (max 60 characters).
   */
  seoTitle?: LocaleString

  /**
   * SEO description — `localeString`
   *
   * Displayed on Facebook and Twitter shares (max 65 characters).
   */
  seoDescription?: LocaleString

  /**
   * SEO Image — `image`
   *
   * Ideal size 1200 x 630px (if not added main image will be used).
   */
  seoImage?: {
    _type: "image"
    asset: SanityReference<SanityImageAsset>
    crop?: SanityImageCrop
    hotspot?: SanityImageHotspot
  }
}

/**
 * Site
 *
 *
 */
export interface Site extends SanityDocument {
  _type: "site"

  /**
   * Site Name — `localeString`
   *
   *
   */
  siteName?: LocaleString

  /**
   * Site Description — `localeString`
   *
   *
   */
  siteDescription?: LocaleString

  /**
   * Keywords — `localeString`
   *
   * A list of keywords seperated by commas.
   */
  keywords?: LocaleString

  /**
   * SEO title — `localeString`
   *
   * Displayed on Facebook and Twitter shares (max 60 characters).
   */
  seoTitle?: LocaleString

  /**
   * SEO description — `localeString`
   *
   * Displayed on Facebook and Twitter shares (max 65 characters).
   */
  seoDescription?: LocaleString

  /**
   * Twitter Handle — `string`
   *
   *
   */
  twitterHandle?: string

  /**
   * SEO Image — `image`
   *
   * Ideal size 1200 x 630px.
   */
  seoImage?: {
    _type: "image"
    asset: SanityReference<SanityImageAsset>
    crop?: SanityImageCrop
    hotspot?: SanityImageHotspot
  }

  /**
   * Opening Heading — `localeString`
   *
   * Heading before opening times (ie Opening Times)
   */
  openingHeading?: LocaleString

  /**
   * Opening Times — `localeString`
   *
   * One line only (ie Thursday - Sunday 11-4)
   */
  openingTimes?: LocaleString

  /**
   * Address Line 1 — `string`
   *
   *
   */
  addressLine1?: string

  /**
   * Address Line 2 — `string`
   *
   *
   */
  addressLine2?: string

  /**
   * Telephone — `string`
   *
   *
   */
  telephone?: string

  /**
   * Email — `string`
   *
   *
   */
  email?: string

  /**
   * Sign Up Title — `localeString`
   *
   * One line only (ie Sign up to our mailing list)
   */
  signUp?: LocaleString

  /**
   * Sign Up Placeholder — `localeString`
   *
   * One line only (ie Email address...)
   */
  signUpPlaceholder?: LocaleString

  /**
   * Social Links — `array`
   *
   *
   */
  socialLinks?: Array<SanityKeyedReference<Social>>
}

/**
 * Social Links
 *
 *
 */
export interface Social extends SanityDocument {
  _type: "social"

  /**
   * site — `string`
   *
   *
   */
  site?: string

  /**
   * Link — `url`
   *
   *
   */
  link?: string
}
// ...
export type CaptionImage = {
  _type: "image"
  asset: SanityReference<SanityImageAsset>
  crop?: SanityImageCrop
  hotspot?: SanityImageHotspot

  /**
   * Caption — `string`
   *
   *
   */
  caption?: string

  /**
   * Attribution — `string`
   *
   *
   */
  Caption?: string
}

export type LocaleRichText = {
  _type: "localeRichText"
  /**
   * English — `blockContent`
   *
   *
   */
  en?: BlockContent

  /**
   * Welsh — `blockContent`
   *
   *
   */
  cy?: BlockContent
}

export type LocaleSlug = {
  _type: "localeSlug"
  /**
   * English — `slug`
   *
   *
   */
  en?: { _type: "en"; current: string }

  /**
   * Welsh — `slug`
   *
   *
   */
  cy?: { _type: "cy"; current: string }
}

export type LocaleString = {
  _type: "localeString"
  /**
   * English — `string`
   *
   *
   */
  en?: string

  /**
   * Welsh — `string`
   *
   *
   */
  cy?: string
}